Separate Constraint with ConstraintDat
This splits the Constraint into a main piece containing the context and the dat
containing the actual data. The purpose of the separation is that the
constraintDat is created locally while the context around it is added later.
This allows the context to contain the results of encoding the ObjArr, and apply
that to the constraints within that same ObjArr.