Adds names to TypeEnv
The names that may be found in a value include not just major names, but also
those for arguments. This will now track all of that so relativeNames can be
correctly expanded. It also starts making the behavior of TypeEnv closer to what
it will eventually be with refinement types and as the classGraph is implemented
with them.