Create DocShow
This creates a new class of DShow that can be a standard represenation of all
values that can be shown in webdocs. Then, the annots can be unified by trying
to show them using dshow.

It also improves the evalAnnots in several ways. They are safely computed
allowing webdocs to function even with errors. It also is run only in the case
that the reaches check suggests it should be doable.