Begins splitting executeConstraint
Right now, executeConstraint is a monolith that is hard to read and understand.
This starts work towards splitting it into three sub-parts: load, compute, and
save. Each part is much simpler and smaller in scope. This commit is copying
pieces from the original and keeping the ones that run without bugs. The ones
requiring more work or having bugs are left with their original
executeConstraint implementation for this commit, and will be fixed in a later
one.