Use ObjArr inside Context
This helps progress unification of the syntax around the same elements. It also
removes more of the remaining usages of types in the formatter (as opposed to
only expressions). This may need to be augmented with additional data, but
should be supplementary to the ObjArr rather than a whole replacement.