Parses all ctx files
This finalizes work to ensure parsing for ctx files. It now adds them to the
Makefile to ensure they remain up to date. It also means the ctx can undergo
syntax migrations using the formatter.

Also has code to handle parsing ro and covered using contexts. However, only
parsing is done and immediately returns an error until it is implemented.