Rewrite buildArrArgs with exprArgsWithVal
The earlier version of buildArrArgs was directed by the value making it both
hacky and have bugs when dealing with some types of expressions. The later
version is simpler and also better aligns with the other exprVarArgs style
functions.

This then fixes the disabled tests for match, case, and ifThenElse.