Converts module to syntax pre-processor
This uses a syntax pre-processor to implement modules. It means the syntax is
more uniform, and it also represents progress towards moving the class system
into the language itself.

As part of this, I also removed the rawExprStatement. Instead, I just give it
special handling where necessary.

It also moves the desugar processing to be semi-recursive. It can now return
arbitrary statements which are passed back into the desugar to run through from
the parsing statement point on.