Makes decDecl recursive
This converts the decl desugar to act recursively through desStatement. One
benefit is that this helps simplify some of the code and removes the
splitDeclSubStatements.

The major benefit, however, is that this opens up support for nested
pre-processors. This includes match inside match or nestedDeclarations inside
other nestedDeclarations. And, it would be a pre-requisite for the bindStatement
in context.ct.