Removes final apply keyword
This moves the apply statement to a combination of a new raw expression type and
a declaration pre-processor. The expression type is written a"terms", somewhat
similar to the python format string. It be extended to hold it like f"" or other
types of format strings.

With this, all of the keywords and restricted terms in the language have been
moved into declaration pre-processors and other types of built-ins.