Removes preprocessor syntax statements
This takes the former statements that are now implemented by the preprocessors
and moves them out of the syntax. This cleans up the formatter and can be
applied to webdocs as well.

As part of this, I also moved the raw expression building utilities to a
separate file. I should also do one for semantics and main expressions later.