Move RawSpread to part of TupleApply
This migrates the spread operation to be part of applying a tuple, with the
typical syntax of val(..spreadArg). There is also a syntax sugar of val.. that
is preserved.

With this, the spread will be a top-level expression. In objects, it represents
matching all other arguments to a value. In arrows, it is applying all arguments
from one. A common behavior will be .._ in objects which matches all arguments
and then ignores them. Using this, I implement a definition for anonymous tuples
and am working on one for context.