Implement EApp for spread support
This adds the support for tuple apply spread into the semantics. It does so by
creating a new wrapper EApp to be used for TupleApply. For a refactor, this can
also be used to for VarApply. However, it will also mean that VarApply uses a
proper base with base arrow rather than the current system (which is probably a
good change).

Right now, most of the spread code remains undefined but will be used more in
the next commit.