Saves ArrowTo reachesTrees
When viewing the type of an ObjArr arrM result, the standard meta type is a poor
substitute. While it does convey some of the information, only the full
reachesTree fully explains the possibilities. So, it is now computed during the
typechecking and then saved into the metadat. This can then be used for mroe
precise tree building and a better checking of evalTargetMode.