Removes usage of fullDest from arrowDestType
I am removing it because it seems to no longer be necessary. In general, it was
a substitute for better usage of the ReachesTree which is being implemented now.
And, that implementation seems to be sufficient. Overall, removing works so it
clearly isn't necessary and may also be responsible for some bugs.