Reaches works for called functions
This fixes a series of bugs that prevent the typeGraph reaches from working
properly with called functions. First, it simplifies the VarMeta reaches
environment by making show work properly. Before this, show actually didn't show
the types, just stored the results in the metadata. Now, it does keep types
similarly to how decode does.

Following this, the typeGraph support is updated. It now runs into situations
where it will fail without the full recursive reaches, so it is added. As part
of this, a reaches cycle detection was also put in to make it not fail the
typeCheckTests.

As a result of this, the examples with called functions are now able to
determine the appropriate type in the parent function.

It also is now passing on the disabled context.ct.