Creates ContextIn and ContextOut
This takes the existing context and splits it into three: ContextIn, ContextOut,
and Context. The ContextIn and ContextOut are used for change, and are what the
desugared format translate into. The main Context is used to represent data, and
the difference between the two is bridged by the new Context macro.

This also creates a new arrExists macro which was partially implemented in some
previous commits, but fully implemented here. It is used to typecheck with the
new context macro definition.