Re-implements evalTargetMode using typeGraph
With the old implementation, evalTargetMode could only work if the evaluation
had a main object. With the new version, it could work even if there wasn't a
specific oa that just matches. It is aimed mostly at making a working Context
system when splitting the Context into ContextIn and ContextOut.