Use names for runtime primitives
This makes the standard runtime primitives be attached to declarations by using
a lookup name. The benefit of this is that it enforces the use of code
declaration (with the opportunity for accompanying documentation and easier
viewability). It also means that the signatures don't have to be written in the
runtime, which makes it much easier to support more complicated signatures that
might be found with macros.