Minor changes for non-prefix type vars
One goal is to remove the type var prefix $ such that $T becomes just T. This is
some changes and fixes in the progress towards the goal. However, it is blocked
by difficulties recognizing type variables when semiDesExpr is called in the
parser. To make it work, we really need to just remove that dependency.