Removes RawObjArr arr Meta for Expr
This finalizes the usage of Expressions for representing metadata in the RawPrgm
by removing the last holdout in the RawArrM.

To complete this change, some of the functions such as exprAppliedArgs and those
using it can't be implemented with RawExpr because they use ObjArr instead of
RawObjArr. Instead, a raw variation are used instead.

The result is that this enables the weird dependency of some parts of parsing on
Desugarf.Expr. It also means that the rest of that expression desugaring is done
with the possibility for having a complete vaenv for detecting references to
vars.