Moves VarApply to EAppVar
This re-implements the existing VarApply logic in terms of TupleApply with
EAppVar. This combines much of the base logic between the two. It also makes
progress towards eventually unifying the args and vars.

However, this also makes a fundamental change in how the vars are implemented.
Before, they were assuming that there was no arrow and the base expression and
meta were equal. The var was just clarifying something about a type that was
already there. This makes it so that there is now a possibility for an arrow
between the base and var application.