Improvements from working towards EAppVar
The goal is to change VarApply into an EAppVar. In the process, here are a few
small improvements.

argToType can support any argument, and it will also break TopTypes that are not PTopType.

Better debugging for during constraint when it is unable to break the
unionAllObjs, so I don't have to go on a goose chase to figure out that is the problem.