Optimize docs TypeInfer by splitting into pages
Before, the TypeInfer would acquire the constrain trace all at once. However,
this can be far too large for any reasonable size file. On one test, it was >.5
GB before I killed it. So, I split it such that the trace half of the page will
query the server for each pnt of data as needed. In the future, it could be
improved with caching. But, this should mean the page is usable again.