Supports macros in TCCond
This makes the cond tree support macros. First, it needs to handle cases where
no else arrow is given. This change essentially removes the test, so it makes it
possible in runtime to determine that the arrows are insufficient. This
verification should be added back later.

It also begins attaching the oa to TCPrim and TCMacro.