Refactor TreeBuild and conditions
This is a significant refactor to clean up TreeBuild. It begins by both
simplifying the existing ResArrowBuildItem and makes it used throughout the
buildCallTree functions. Most of the logic can now be observed by following
buildPartialCallTree (previously envLookup). It also renames the functions to
make them clear in behavior. Most can be understood by following the comments.

One change worth noting is the addition of wrapCErr. This is a very useful
utility for the CErr system because it allows a clearer stack trace without
propagating the information all the way down to the error. Errors can be
contextualized higher up the stack when an error is reached further down. This
is used initially in TreeBuild, but can be applied to more places later.