Removes completeTreeSet check with conds
This makes the completeTreeSet system no longer validate completeness when using
conditions. The main issue is refinment types, but the specific one requiring
this slightly different workaround is having ?-> usages. This looks like a
condition, but is actually always True. So it looks like a condition but is
actually a complete set in itself. Without refinement types or a system capable
of recognizing this, I can only disable the check to move forwards.