Improvements to TypeGraph
This is a few improvements to the typegraph. First, the printing is improved
with a custom show using String builder. It also adds a reason to it by creating typeGraphQueryWithReason.

Lastly, it moves the reachesPartial into a reachesPartials. This is because a
tree can't be built without a root, which is represented by the partial type.
This makes the organization around that work out a bit cleaner.