Cleans up arrowConstrainUbs
This simplifies how it was handling the reaches, including using the new
reachesPartials instead of rootReachesPartial. It also abandons the pretense of
filtering the src, although this may be added back later.

As part of this, it also has to clean up unionReachesTree, including by adding
the vaenv to it.