[WD] Migrate from create-react-app to vite
This migrates the base template of the webdocs system from one based on
create-react-app to one based on vite. The main reason is the create-react-app
and the react-scripts it is using seem abandoned. Instead, vite seems more
active and uses newer dependencies for a hopefully better dev server and build.
It also starts opening up the way to using typescript.