Adds initial LSP and vscode extension
This begins the process of creating editor integration. There is a sample of an
LSP server alongside a vscode extension using the server. The vscode
implementation shouldn't need much changes for a while as it relies on
capabilities reported by the server. The server is filled with dummy
implementation so it requires significant changes.