Add initial SemanticTokens to LSP
This augments the LSP with semantic tokens and a simple model of compilation. It
will recompute everything, so it has no sense of incremental computation. It is
built on the existing webdocs system, which probably needs to be refactored out
and expanded.

There are also a few issues. It won't register catln files outside of a loaded
folder. It also breaks by being unable to find the core.