Converts the FileImport to obj
This uses the existing RawFileImport object and extends it to apply to the
semantics import as well. This allows for more data within the main import. It
also serves as progress towards better unifying the import and lookup system to
always use semantics rather than raw imports (such as in webdocs).