Split golden integration by file
Before, the golden tests saved the entire prgmGraph, resulting in significant
duplicates of the common files (core). This divides the file structure so that
the core files are combined into the same place. This reduces the size and
complexity of the golden files, and will reduce commit size when changing them.