Use FileImport for prgmGraph
This makes all program graphs, including the raw one, use FileImport instead of
RawFileImport. It also cleans up some of the parser logic by ensuring that the
FileImport is canonical and the RawFileImport is not.

The goal is that it will make it easier to do logic and conversions for the
entire program pipeline, including from the RawPrgm level.