Moves pipeline logic to CtService
This starts combining much of the pipeline logic previously split into different
files into a common CtService. This is planned to eventually include pieces such
as incremental build and downloading of catln core libraries. It also helps
unify logic so the steps can be better combined and made more resilient.