Simplify with CtService helpers
This adds a number of CtService helpers to simplify the usage and retrieval of
data. Importantly, it adds a unified ctssGet that can be used to add the JIT
building for incremental builds.

Of some note, it also has a combined ctssGetJoined that uses the recently added
Monoid instance for Prgm.