Convert CTSS to use MVar
This makes the CTSS store data inside of an MVar. This simplifies much of the
interaction around reading and building, often down to a single setup command.
At this point, incremental building should be fully working. What is left to
change is to add partial building on the various operations that don't require
the full buildAll.