Use CTSSConfig for stack paths
This enables a new CTSSConfig that can differentiate between a testing local
path for the stack and a required global one to ~/.stack. The resutl is that the
catln executable can now be run from any directory, not just the core of a
cloned catln repo. It does still require setting up and having a cloned repo.
This may get fixed later with releases.