Convert Object Map to map
Originally, the object map was a list. However, it is more efficient to
instead keep it as a map to avoid repeated conversions into a TypeGraph.
Additionally, it should also make it easier to manage the interactions
between different ObjArr that are combined and ensure things like only a
single declaration or that all definitions fall into it or to separate
refinement declarations.