Check declarations and definitions are matching
This adds an additional verification to the decoding phase that will
verify all of the object map items for both definitions outside of
declarations and unfulfilled declarations. This is a key user facing
validation to ensure the code is valid. It also helps during compiler
development by preventing bad test code.